Move MpdContext out of dispatcher module
