+++ src/mopidy_mpd/protocol/current_playlist.py
@@ -8,7 +8,7 @@ from mopidy_mpd import exceptions, protocol, translator
 if TYPE_CHECKING:
     from mopidy.types import Uri
 
-    from mopidy_mpd.dispatcher import MpdContext
+    from mopidy_mpd.context import MpdContext
 
 
 @protocol.commands.add("add")
